Pre-0.8.4 compatibility.
0.8.4 introduced an incompatibility with WebTest as WebTest subclasses
WSGIServer as webtest.http.StopableWSGIServer. This commit introduces
changes to make WSGIServer subclassable once more by making the
signature of BaseWSGIServer's __init__ method compatible with the
pre-0.8.4 WSGIServer class, and by making WSGIServer an alias of
TcpWSGIServer.

This method was suggested and chosen from a discussion with Laurence
Rowe (@lrowe).