Workaround a couple of bugs in OS X
Add a workaround for issues with getaddrinfo() on OS X when using a
.local address based hostname for the listen argument.

i.e. Sterling.local would return not just return a list of addresses,
that list of addresses may contain duplicates. On top of that, OS X also
returns the link-local address multiple times, yet with different zone
identifiers, however upon calling bind() on those it would happily blow
up.

This is definitely unexpected behaviour, so we work around it here.