Set IPv6 socket to be IPv6 only
On systems that allow for IPv6 mapped IPv4 we want to make sure we
explicitly set the socket to be IPv6 only. This way we don't run into
the issue whereby we try to bind to the same socket twice (once with an
mapped IPv6, and then on IPv4) because bind() won't like that.