On Windows IPPROTO_IPV6 may not be defined in socket
Unfortunately due to some reshuffling of headers on Windows, and Python
defining the wrong Windows version, IPPROTO_IPV6 may not be defined on
Windows. If this is the case, we set them to some magic values.

If there is an OS out there that doesn't define those flags, and they
are not windows, we raise a RuntimeError because we need to add some
magic variables, waitress requires IPv6...