Change get_server_name logic to be more correct
In case of wildcard INADDR_ANY or INADDR6_ANY, we try and get the
hostname of the machine we are running on, since it is arguably the most
correct response for a daemon listening on all addresses.

However if that fails, we fall back to returning `localhost`.

If we have an IP address other than the wildcard, we try to get the
hostname by doing a reverse DNS lookup, if this fails we just set the
server_name to the IP address directly.

For IPv6, we MUST wrap the IP address in `[]` so that it is a valid IPv6
literal address.

We also catch the UnicodeDecodeError exception and make sure to ignore
it, this is due to Windows and gethostbyaddr doing the wrong thing
internally.

Closes #186, #201, #168