added any as valid value for the trusted_proxy adjustable
Backgound is when waitress is running inside a container, any
request is originates from the gateway of the container network, not
from the outside ip. The container network changes every time the
container is restarted to a random new network number. So setting one
IP as trusted_proxy isn't possible but still needed as most of the time
an reverse proxy is between the client an the container. So if it's
ensured that only requests from the reverse proxy can reach the
container, trusting any host is not a problem.