Allow requests with URI that starts with multiple slashes
The HTTP spec states that it is acceptable to send a request like:

GET //whatever/testing HTTP/1.1

This should get properly conveyed to the WSGI application, but due to
the way that urlsplit works in the standard library this did not happen
correctly. With this fix we pass through the original path as requested
by the client, and the WSGI application will be responsible for
collapsing multiple empty path segments as necessary.

Fixes #260