Upon processing errors, set Connection: close header
When we start sending out an exception that has happened we want to drop
any further pipelined requests that may be waiting. The way we do this
is by slamming the connection shut after we send the response to the
received request that has errored.

This avoids potential issues with attempting to parse the rest of the
data that is in the TCP/IP connection that may contain invalid data.