Drop Python 3.4
This is necessary due to needing coverage>=5.0, which is no longer
available on Python 3.4.

Python 3.4 is also now past end of life.

This continued to work by chance due to Travis having nosetest installed
in the Python global installation, and thus the failure from tox to
install coverage>=5.0 did not cause errors on Travis.

Testing locally I still had an old version of coverage installed, and
that version had not yet dropped Python 3.4.