Stop marking socket as readable when flushing data
We no longer mark the socket as readable if we are attempting to flush
whatever remaining data we have and are trying to shut down the channel.

Whatever data is ready to be read, it's no longer our concern.

We don't want to spend time reading data we don't care about.