Bugfix on Windows: OSError is not subscriptable
When Waitress fails to launch on Windows due to an issue with the
trigger socket not being ready for connections, we attempt to loop. In
the past this was done by subscripting the OSError and checking to see
if it matched errno.WSAEADDRINUSE, this is no longer possible in newer
verisons of Python.

This is a quick bugfix for a rare case which should no longer happen on
Windows.