Make use of ident while showing server error (generated by waitress)
this will help users to hide the server details from waitress generated error response with the help of ident value.
Some organization don't want to expose there server details to user, since it will help the attackers perform attacks based on the known vulnerability of the server.
So instead of showing generated by waitress, we can show a generic text.