Setting socket options may fail if remote has already disappeared
By the time we come around to accepting a connection on macOS, the
remote may have already disappeared, in which case setting options on
the socket may fail with an error.

Instead of failing and shutting down the server we want to continue, so
move the setting of socket options into the try/except block and log the
error instead.