Lean on pkgutil.resolve_name for importing apps
With #445, it no longer makes sense not to use `pkgutil.resolve_name` to
handle the importing of WSGI applications in the runner. There is one
very minor behavioural change in that `pkgutil.resolve_name` allows a
broader syntax for the imports, but supports the same syntax as the
existing code being removed, so this is a minor concern.

It has the pleasant side effect of removing some legacy code only
present for Python 2.6 that was missed when Python 2 support was
removed.