If Transfer-Encoding includes "chunked" multiple times, drop request
Waitress only supports a single transfer-encoding, that is chunked, and
if it received a request with multiple chunked transfer encodings
requested, it would dutifully unwrap the first layer, turn it into a
request with a content-length for the WSGI application, potentially
leaving the body with incorrect or bad data.

Instead we now drop the request when we receive the request to have
multiple chunked transfer encodings.