When Transfer-Encoding is set, and Content-Length, close connection
RFC9112 adds an additional requirement when parsing a request that has
both content-length and transfer-encoding set, which is that the
connection should be closed after a single use to avoid potential
pitfalls with request smuggling/parser desync between multiple
implementations.

So now we check to see if the content-length was also set, and if so, we
close the connection after the request is completed.